img_resizer/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.11.0
idna==2.8
libmagic==1.0
pilkit==2.0
Pillow==7.0.0
python-magic==0.4.15
python-memcached==1.59
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
